Moved all the conversion methods to their own static class
