+++ iiwa_ros_java/src/de/tum/in/camp/kuka/ros/MessageGenerator.java
@@ -23,9 +23,6 @@
 
 package de.tum.in.camp.kuka.ros;
 
-import geometry_msgs.Pose;
-import geometry_msgs.Quaternion;
-
 import java.util.Arrays;
 
 import org.ros.message.MessageFactory;
@@ -33,15 +30,9 @@ import org.ros.node.NodeConfiguration;
 import org.ros.time.TimeProvider;
 
 import com.kuka.connectivity.motionModel.smartServo.SmartServo;
-import com.kuka.roboticsAPI.deviceModel.JointPosition;
 import com.kuka.roboticsAPI.deviceModel.LBR;
-import com.kuka.roboticsAPI.geometricModel.Frame;
 import com.kuka.roboticsAPI.geometricModel.ObjectFrame;
-import com.kuka.roboticsAPI.geometricModel.math.Matrix;
-import com.kuka.roboticsAPI.geometricModel.math.MatrixBuilder;
-import com.kuka.roboticsAPI.geometricModel.math.MatrixRotation;
 import com.kuka.roboticsAPI.geometricModel.math.Transformation;
-import com.kuka.roboticsAPI.geometricModel.math.Vector;
 import com.kuka.roboticsAPI.motionModel.controlModeModel.JointImpedanceControlMode;
 
 /**
@@ -49,7 +40,7 @@ import com.kuka.roboticsAPI.motionModel.controlModeModel.JointImpedanceControlMo
  * it's a collection of methods to build the messages from the current state of a LBR iiwa Robot.
  * For Cartesian messages, it's possible to pass a reference frames, if no reference frames is passed, the flange frame is used.
  */
-public class iiwaMessageGenerator {
+public class MessageGenerator {
 
 	private static String baseFrameID;
 	private static final String baseFrameIDSuffix = "_link_0";
@@ -61,9 +52,9 @@ public class iiwaMessageGenerator {
 	// Objects to create ROS messages
 	private NodeConfiguration nodeConf = NodeConfiguration.newPrivate();
 	private MessageFactory messageFactory = nodeConf.getTopicMessageFactory();
-	private TimeProvider time = iiwaConfiguration.getTimeProvider();
+	private TimeProvider time = Configuration.getTimeProvider();
 
-	public iiwaMessageGenerator(String robotName) {
+	public MessageGenerator(String robotName) {
 		baseFrameID = robotName + baseFrameIDSuffix; // e.g. if robotName == iiwa, then baseFrameID = iiwa_link_0
 
 		// e.g. if robotName == iiwa, the joints are iiwa_joint_1, iiwa_joint_2, ...
@@ -103,7 +94,7 @@ public class iiwaMessageGenerator {
 		currentPose.getHeader().setFrameId(baseFrameID);  
 		currentPose.getHeader().setStamp(time.getCurrentTime());
 
-		kukaTransformationToRosPose(transform, currentPose.getPose());
+		Conversions.kukaTransformationToRosPose(transform, currentPose.getPose());
 	}
 
 
@@ -153,7 +144,7 @@ public class iiwaMessageGenerator {
 	public void getCurrentJointPosition(iiwa_msgs.JointPosition currentJointPosition, LBR robot) {
 		double[] position = robot.getCurrentJointPosition().getInternalArray();
 		currentJointPosition.getHeader().setStamp(time.getCurrentTime());
-		vectorToJointQuantity(position, currentJointPosition.getPosition());
+		Conversions.vectorToJointQuantity(position, currentJointPosition.getPosition());
 	}
 
 	/**
@@ -166,8 +157,8 @@ public class iiwaMessageGenerator {
 		double[] position = robot.getCurrentJointPosition().getInternalArray();
 		currentJointPositionVelocity.getHeader().setStamp(time.getCurrentTime());
 
-		vectorToJointQuantity(position, currentJointPositionVelocity.getPosition());
-		vectorToJointQuantity(computeVelocity(robot), currentJointPositionVelocity.getVelocity());
+		Conversions.vectorToJointQuantity(position, currentJointPositionVelocity.getPosition());
+		Conversions.vectorToJointQuantity(computeVelocity(robot), currentJointPositionVelocity.getVelocity());
 	}
 
 	/**
@@ -176,9 +167,10 @@ public class iiwaMessageGenerator {
 	 * @param robot : an iiwa Robot, its current state is used to set the values of the message.
 	 */
 	public void getCurrentJointVelocity(iiwa_msgs.JointVelocity currentJointVelocity, LBR robot) {
-		vectorToJointQuantity(computeVelocity(robot), currentJointVelocity.getVelocity());
+		Conversions.vectorToJointQuantity(computeVelocity(robot), currentJointVelocity.getVelocity());
 	}
 
+	// TODO: doc
 	private double[] computeVelocity(LBR robot) {
 		double[] position = robot.getCurrentJointPosition().getInternalArray();
 		long position_time_ns = System.nanoTime();
@@ -215,7 +207,7 @@ public class iiwaMessageGenerator {
 		}
 
 		currentJointStiffness.getHeader().setStamp(time.getCurrentTime());
-		vectorToJointQuantity(stiffness, currentJointStiffness.getStiffness());
+		Conversions.vectorToJointQuantity(stiffness, currentJointStiffness.getStiffness());
 	}
 
 	/**
@@ -239,7 +231,7 @@ public class iiwaMessageGenerator {
 		}
 
 		currentJointDamping.getHeader().setStamp(time.getCurrentTime());
-		vectorToJointQuantity(damping, currentJointDamping.getDamping());
+		Conversions.vectorToJointQuantity(damping, currentJointDamping.getDamping());
 	}
 
 	/**
@@ -251,7 +243,7 @@ public class iiwaMessageGenerator {
 	public void getCurrentJointTorque(iiwa_msgs.JointTorque currentJointTorque, LBR robot) {
 		double[] torque = robot.getMeasuredTorque().getTorqueValues();
 		currentJointTorque.getHeader().setStamp(time.getCurrentTime());
-		vectorToJointQuantity(torque, currentJointTorque.getTorque());		
+		Conversions.vectorToJointQuantity(torque, currentJointTorque.getTorque());		
 	}
 
 	/**
@@ -266,246 +258,6 @@ public class iiwaMessageGenerator {
 		currentJointState.setName(Arrays.asList(joint_names));
 		currentJointState.setPosition(robot.getCurrentJointPosition().getInternalArray());
 		currentJointState.setEffort(robot.getMeasuredTorque().getTorqueValues());
-	}
-
-	// Conversions
-
-	/**
-	 * Generates a iiwa_msgs.JointQuantity message from a double vector
-	 * @param vec : double vector
-	 * @param q : resulting JointQuantity message
-	 */
-	public void vectorToJointQuantity(double[] vec, iiwa_msgs.JointQuantity q) {
-		q.setA1((float) vec[0]);
-		q.setA2((float) vec[1]);
-		q.setA3((float) vec[2]);
-		q.setA4((float) vec[3]);
-		q.setA5((float) vec[4]);
-		q.setA6((float) vec[5]);
-		q.setA7((float) vec[6]);
-	}
-
-	/**
-	 * Generate a double vector from a iiwa_msgs.JointQuantity message
-	 * @param q : a JointQuantity message
-	 * @return a double vector
-	 */
-	public double[]  jointQuantityToVector(iiwa_msgs.JointQuantity q) {
-		double[] ret = new double[7];
-
-		ret[0] = q.getA1();
-		ret[1] = q.getA2();
-		ret[2] = q.getA3();
-		ret[3] = q.getA4();
-		ret[4] = q.getA5();
-		ret[5] = q.getA6();
-		ret[6] = q.getA7();
-
-		return ret;
-	}
-
-	/**
-	 * Generates a MatrixRotation from a Quaternion
-	 * @param x
-	 * @param y
-	 * @param z
-	 * @param w
-	 * @return a MatrixRotation
-	 */
-	public static MatrixRotation quatToMatrix(double x, double y, double z, double w) {
-		return quatToMatrix((float) x, (float) y, (float) z, (float) w);
-	}
-
-	/**
-	 * Generates a MatrixRotation from a Quaternion
-	 * @param x
-	 * @param y
-	 * @param z
-	 * @param w
-	 * @return a MatrixRotation
-	 */
-	public final static MatrixRotation quatToMatrix(float x, float y, float z, float w) {
-		double sqw = w*w;
-		double sqx = x*x;
-		double sqy = y*y;
-		double sqz = z*z;
-
-		MatrixBuilder mb = new MatrixBuilder();
-
-		// invs (inverse square length) is only required if quaternion is not already normalised
-		double invs = 1 / (sqx + sqy + sqz + sqw);
-		mb.setElement00(( sqx - sqy - sqz + sqw)*invs) ; // since sqw + sqx + sqy + sqz =1/invs*invs
-		mb.setElement11((-sqx + sqy - sqz + sqw)*invs);
-		mb.setElement22((-sqx - sqy + sqz + sqw)*invs);
-
-		double tmp1 = x*y;
-		double tmp2 = z*w;
-		mb.setElement10(2.0 * (tmp1 + tmp2)*invs);
-		mb.setElement01(2.0 * (tmp1 - tmp2)*invs);
-
-		tmp1 = x*z;
-		tmp2 = y*w;
-		mb.setElement20(2.0 * (tmp1 - tmp2)*invs);
-		mb.setElement02(2.0 * (tmp1 + tmp2)*invs);
-
-		tmp1 = y*z;
-		tmp2 = x*w;
-		mb.setElement21(2.0 * (tmp1 + tmp2)*invs);
-		mb.setElement12(2.0 * (tmp1 - tmp2)*invs);
-
-		return MatrixRotation.of(mb.toMatrix());
-	}
-
-	/**
-	 * Generates a quaternion from a Matrix.<p>
-	 * Mercilessly copied <url>from https://github.com/libgdx/libgdx/blob/master/gdx/src/com/badlogic/gdx/math/Quaternion.java</url>
-	 * @param matrix : the starting matrix
-	 * @param quaternion : the resulting quaternion
-	 */
-	public void matrixToQuat(Matrix matrix, Quaternion quaternion) {
-
-		double xx = matrix.getElement00();
-		double xy = matrix.getElement01();
-		double xz = matrix.getElement02();
-		double yx = matrix.getElement10();
-		double yy = matrix.getElement11();
-		double yz = matrix.getElement12();
-		double zx = matrix.getElement20();
-		double zy = matrix.getElement21();
-		double zz = matrix.getElement22();
-
-		double x,y,z,w; // return
-
-		final double t = xx + yy + zz;
-
-		// we protect the division by s by ensuring that s>=1
-		if (t >= 0) { // |w| >= .5
-			float s = (float)Math.sqrt(t + 1); // |s|>=1 ...
-			w = 0.5f * s;
-			s = 0.5f / s; // so this division isn't bad
-			x = (zy - yz) * s;
-			y = (xz - zx) * s;
-			z = (yx - xy) * s;
-		} else if ((xx > yy) && (xx > zz)) {
-			float s = (float)Math.sqrt(1.0 + xx - yy - zz); // |s|>=1
-			x = s * 0.5f; // |x| >= .5
-			s = 0.5f / s;
-			y = (yx + xy) * s;
-			z = (xz + zx) * s;
-			w = (zy - yz) * s;
-		} else if (yy > zz) {
-			float s = (float)Math.sqrt(1.0 + yy - xx - zz); // |s|>=1
-			y = s * 0.5f; // |y| >= .5
-			s = 0.5f / s;
-			x = (yx + xy) * s;
-			z = (zy + yz) * s;
-			w = (xz - zx) * s;
-		} else {
-			float s = (float)Math.sqrt(1.0 + zz - xx - yy); // |s|>=1
-			z = s * 0.5f; // |z| >= .5
-			s = 0.5f / s;
-			x = (xz + zx) * s;
-			y = (zy + yz) * s;
-			w = (yx - xy) * s;
-		}
-
-		quaternion.setX(x);
-		quaternion.setY(y);
-		quaternion.setZ(z);
-		quaternion.setW(w);
-	}
-
-	/**
-	 * Converts a geometry_msgs.Pose message to a Transformation object in KUKA APIs
-	 * @param rosPose : starting Pose
-	 * @return resulting Transformation
-	 */
-	public Transformation rosPoseToKukaTransformation(geometry_msgs.Pose rosPose) {
-		if (rosPose == null)
-			return null;
-
-		float tx = (float) rosPose.getPosition().getX()*1000;
-		float ty = (float) rosPose.getPosition().getY()*1000;
-		float tz = (float) rosPose.getPosition().getZ()*1000;
-
-		float x = (float) rosPose.getOrientation().getX();
-		float y = (float) rosPose.getOrientation().getY();
-		float z = (float) rosPose.getOrientation().getZ();
-		float w = (float) rosPose.getOrientation().getW();
-
-		MatrixRotation rot = iiwaMessageGenerator.quatToMatrix(x, y, z, w);
-		Vector transl = Vector.of(tx, ty, tz);
-
-		return Transformation.of(transl, rot);
-	}
-
-	/**
-	 * Converts a geometry_msgs.Pose message to a Frame object in KUKA APIs
-	 * @param rosPose : starting Pose
-	 * @return resulting Frame
-	 */
-	public Frame rosPoseToKukaFrame(geometry_msgs.Pose rosPose) {	
-		return new Frame(rosPoseToKukaTransformation(rosPose));
-	}
-
-	/**
-	 * Converts a Transformation object from KUKA APIs to a geometry_msgs.Pose message
-	 * @param kukaTransf : starting Trasnformation
-	 * @param pose : resulting Pose
-	 */
-	public void kukaTransformationToRosPose(Transformation kukaTransf, Pose pose) {
-		pose.getPosition().setX(kukaTransf.getX()/1000); 
-		pose.getPosition().setY(kukaTransf.getY()/1000);
-		pose.getPosition().setZ(kukaTransf.getZ()/1000);
-
-		Matrix rotationMatrix = kukaTransf.getRotationMatrix();
-		matrixToQuat(rotationMatrix, pose.getOrientation());
-	}
-
-	/**
-	 * Converts an iiwa_msgs.JointQuantity to a JointPosition in KUKA APIs
-	 * @param rosJointPos : starting JointQuantity
-	 * @param kukaJointPos : resulting JointPosition
-	 */
-	public void rosJointQuantityToKuka(iiwa_msgs.JointQuantity rosJointPos, JointPosition kukaJointPos) {
-		rosJointQuantityToKuka(rosJointPos, kukaJointPos, 1.0);
-	}
-
-	/**
-	 * Converts an iiwa_msgs.JointQuantity to a JointPosition in KUKA APIs
-	 * @param rosJointPos : starting JointQuantity
-	 * @param kukaJointPos : resulting JointPosition
-	 * @param scaleFactor : each element of rosJointPos with be scaled using this value 
-	 */
-	public void rosJointQuantityToKuka(iiwa_msgs.JointQuantity rosJointPos, JointPosition kukaJointPos, double scaleFactor) {
-		kukaJointPos.set(
-				rosJointPos.getA1()*scaleFactor,
-				rosJointPos.getA2()*scaleFactor,
-				rosJointPos.getA3()*scaleFactor,
-				rosJointPos.getA4()*scaleFactor,
-				rosJointPos.getA5()*scaleFactor,
-				rosJointPos.getA6()*scaleFactor,
-				rosJointPos.getA7()*scaleFactor
-				);
-	}
-
-	/**
-	 * Converts an iiwa_msgs.JointQuantity to a double vector.
-	 * @param rosJointQuant : starting jointQuantity
-	 * @return a double vector
-	 */
-	public double[] rosJointQuantityToArray(iiwa_msgs.JointQuantity rosJointQuant) {
-		double[] ret = {
-				rosJointQuant.getA1(),
-				rosJointQuant.getA2(),
-				rosJointQuant.getA3(),
-				rosJointQuant.getA4(),
-				rosJointQuant.getA5(),
-				rosJointQuant.getA6(),
-				rosJointQuant.getA7()
-		};
-		return ret;
-
 	}
 
 	/**
